Implement GET and TRAFFIC jobs

+++ orders-common/src/main/java/org/gangel/orders/job/Configuration.java
@@ -12,14 +12,14 @@ public class Configuration {
     
     public static JobType jobType = JobType.UNKNOWN;
     
-    public static int minCustomerId = 1;
-    public static int maxCustomerId = 10_000;
+    public static long minCustomerId = 1;
+    public static long maxCustomerId = 10_000;
     
-    public static int minProductId = 1;
-    public static int maxProductId = 10_000;
+    public static long minProductId = 1;
+    public static long maxProductId = 10_000;
     
-    public static int minOrdersId = 1;
-    public static int maxOrdersId = 10_000;
+    public static long minOrdersId = 1;
+    public static long maxOrdersId = 10_000;
     
     public static boolean isSSL = false; 
     public static String sslCertFile;
